requirements.txt
Flask_SocketIO==3.3.2
Flask==1.0.2
graphqlclient==0.2.4
pyserial==3.4
